backend/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface
langchain-ollama
python-dotenv
pydantic
pypdf
pymupdf
langchain-text-splitters
sentence-transformers>=2.2.2
faiss-cpu
fastapi
uvicorn
chromadb
python-multipart 
sqlalchemy
transformers
torch